split and refactor test classes
