+++ integration-tests/src/test/kotlin/com/testapp/tests/payments/flows/paylater/PayLaterSETest.kt
@@ -1,0 +1,22 @@
+package com.testapp.tests.payments.flows.paylater
+
+import com.testapp.network.KlarnaApi
+import com.testapp.utils.BillingAddressTestHelper
+import com.testapp.utils.SessionHelper
+import org.junit.Test
+
+internal class PayLaterSETest : BasePayLaterTest() {
+
+    @Test
+    fun `test payment pay later sweden successful flow`() {
+        val session = KlarnaApi.getSessionInfo(SessionHelper.getRequestSE())?.session
+        testPayLater(true, session, BillingAddressTestHelper.getBillingInfoSE())
+    }
+
+    @Test
+    fun `test payment pay later sweden failure flow`() {
+        val session = KlarnaApi.getSessionInfo(SessionHelper.getRequestSE())?.session
+        testPayLater(false, session, BillingAddressTestHelper.getBillingInfoSE())
+    }
+
+}
